# Plugins and Modules

## Changes

- unique errors for connection and client state verification in
ibc-solidity connection handshake verification
- more attributes on generated prost types for `cometbft-rpc`
- more canonical types in `unionlabs` for generated prost types
- the great untyping:
- `beacon-api-client` is now fully generic over chain spec and does not
take a generic for chain spec
- `unionlabs::ics24::Path` is no longer generic, using only concrete
types
- `unionlabs::ics24::IbcPath` is no longer generic, using only concrete
types
- IbcPath::Value for client and consensus state paths are now just
opaque bytes
- `unionlabs::event::IbcEvent` is no longer generic, using only concrete
types
- all of the IBC `Msg*` are no longer generic, using only concrete types
- added `Unbounded*` versions of all of the types in
`unionlabs::ethereum::beacon` that are generic over ChainSpec, for use
in `beacon-api-client`
    - open to better naming here
- `unionlabs::QueryHeight` is no longer generic, using only the concrete
type `Height`
- removed a bunch of old commented out code
- queue_msg:
- removed `OpEvent`/`QueueMessage::Event` (no longer useful with the new
architecture)
- added new `RawAny` type, since some cosmos types use a dynamic list of
`Any` values and as such they must be type-erased
- cleaned up `unionlabs::ethereum::beacon` to follow the structure of
the rest of the crate (one module/file per type)
- `#[model]` now derives `Eq`
- a lot of other stuff I'm too lazy to write out

## Closes

- https://github.com/unionlabs/union/issues/1813 (no longer relevant)

## Non-code refactors

- normalize comments in rust code (mostly affected generated code)

---

`chain-utils` is a bit of a mess right now, this crate will be
drastically refactored in the coming weeks. i recommend just kinda not
looking at it too hard

+++ hubble/src/bera.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 use backon::{ConstantBuilder, ExponentialBuilder, Retryable};
 use color_eyre::{eyre::eyre, Result};
-use cometbft_rpc::{AbciQueryResponse, Client};
+use cometbft_rpc::{types::AbciQueryResponse, Client};
 use tracing::info;
 use unionlabs::{
     berachain::BerachainChainSpec, encoding::DecodeAs,
@@ -62,7 +62,7 @@ impl Bera {
                 "store/beacon/key",
                 data,
                 Some(
-                    (slot - 1)
+                    (slot as i64 - 1)
                         .try_into()
                         .expect("converting slot to abci_query slot"),
                 ),
